\name{survData}
\alias{survData}
\docType{data}
\title{
Survival data
}
\description{
The survival data of patients in \code{mRNA_matrix}.
}
\usage{data("survData")}
\format{
    A data frame with 314 observations on the following 2 variables.
    \describe{
        \item{\code{status}}{a logical vector}
        \item{\code{time}}{a numeric vector}
    }
}
\examples{
data(survData)
}
\keyword{datasets}
